+++ src/routes/envman/storage.ts
@@ -2,12 +2,14 @@ import { Elysia } from 'elysia'
 import { storageCoreRouter } from './storage-core'
 import { storageFolderRouter } from './storage-folder'
 import { storageMoveRouter } from './storage-move'
+import { storagePresignRouter } from './storage-presign'
 import { storageRenameRouter } from './storage-rename'
 import { storageUploadRouter } from './storage-upload'
 
 export const storageRouter = new Elysia()
   .use(storageCoreRouter)
   .use(storageUploadRouter)
+  .use(storagePresignRouter)
   .use(storageRenameRouter)
   .use(storageMoveRouter)
   .use(storageFolderRouter)